+++ asyncdex/__init__.py
@@ -1,5 +1,5 @@
 from .client import MangadexClient
-from .exceptions import AsyncDexException, HTTPException, Ratelimit, Unauthorized
+from .enum import ContentRating, Demographic, FollowStatus, MangaStatus, Relationship, Visibility
+from .exceptions import AsyncDexException, HTTPException, Missing, Ratelimit, Unauthorized
 from .ratelimit import Ratelimits
 from .version import version
-from .enum import ContentRating, Demographic, Relationship, MangaStatus, FollowStatus, Visibility