+++ asyncdex/models/__init__.py
@@ -1,0 +1,3 @@
+from .author import Author
+from .manga import Manga
+from .tag import Tag